Refactor : Create a seperate C file for UART functions
