+++ .gitignore
@@ -1,0 +1,4 @@
+/PSLab_Original/build/default/
+/PSLab_Original/dist/default/
+/MPLABProject/nbproject/private/
+/MPLABProject/build/default/
