--- MPLABProject/nbproject/project.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://www.netbeans.org/ns/project/1">
-    <type>com.microchip.mplab.nbide.embedded.makeproject</type>
-    <configuration>
-        <data xmlns="http://www.netbeans.org/ns/make-project/1">
-            <name>MPLABProject</name>
-            <creation-uuid>5e50493f-2124-4cd0-8a48-5cdfee5b2f55</creation-uuid>
-            <make-project-type>0</make-project-type>
-            <c-extensions>c</c-extensions>
-            <cpp-extensions/>
-            <header-extensions>h</header-extensions>
-            <asminc-extensions/>
-            <sourceEncoding>ISO-8859-1</sourceEncoding>
-            <make-dep-projects/>
-        </data>
-    </configuration>
-</project>
