+++ PSLab_Original/functions.c
@@ -1,4 +1,3 @@
-
 /*
  * File:   newmain.c
  * Author: jithin
@@ -22,7 +21,7 @@
  * 15nS time resolution. 16-bit, 32-bit modes
  */
 
-#include"functions.h"
+#include "functions.h"
 
 void __attribute__((interrupt, no_auto_psv)) _AD1Interrupt(void) {
     _AD1IF = 0;
@@ -114,98 +113,6 @@ void __attribute__((__interrupt__, no_auto_psv)) _INT2Interrupt(void)
     INITIAL_DIGITAL_STATES_ERR=((PORTB>>10)&0xF)|(_C4OUT<<4); // =(PORTB>>10)&0xF;
 
     _INT2IF=0;_INT2IE=0;
-
-}
-
-
-void __attribute__((__interrupt__, no_auto_psv)) _U2RXInterrupt(void)
-{
-    asm("CLRWDT");
-//while (U1STAbits.UTXBF); //wait for transmit buffer empty
-U1TXREG = U2RXREG;
-_U2RXIF=0;
-}
-void __attribute__((__interrupt__, no_auto_psv)) _U1RXInterrupt(void)
-{
-    asm("CLRWDT");
-while (U2STAbits.UTXBF); //wait for transmit buffer empty
-U2TXREG = U1RXREG;
-_U1RXIF=0;
-
-}
-
-
-void configUART2(unsigned int BAUD){
-    _TRISB5=0;_TRISB6=1;
-    RPOR1bits.RP37R=0x03;RPINR19bits.U2RXR = 38;
-
-    U2MODEbits.STSEL = 0; //1 stop bit
-    U2MODEbits.PDSEL = 0; //no parity, 8-data bits
-    U2MODEbits.ABAUD = 0; //disable auto-baud
-    U2MODEbits.BRGH = 1; //high speed mode
-
-    U2BRG = BAUD;
-    
-    U2MODEbits.UEN = 0;
-    U2MODEbits.RTSMD = 1;
-
-    U2STAbits.URXISEL = 0; //interrupt on 1 char recv
-
-    U2MODEbits.UARTEN = 1; //enable UART
-    U2STAbits.UTXEN = 1; //enable UART TX
-
-    U2MODEbits.URXINV = 0;
-    while(U2STAbits.URXDA)U2RXREG;
-    
-}
-
-bool hasChar2(void) {
-    return U2STAbits.URXDA;
-}
-
-char getChar2(void) {
-    if (!hasChar2())return 0;
-    if (U2STAbits.FERR == 1) {
-        U2RXREG;
-        U2STAbits.OERR = 0;
-        U2STAbits.FERR = 0;
-        return 0;
-    }
-    return U2RXREG;
-}
-
-unsigned int getInt2(void) {
-    c1 = getChar2()&0xFF;
-    c2 = getChar2()&0xFF;
-    return (c2 << 8) | c1;
-}
-
-void sendChar2(char val) {
-    while (U2STAbits.UTXBF); //wait for transmit buffer empty
-    U2TXREG = val;
-}
-
-void sendInt2(unsigned int val) {
-    while (U2STAbits.UTXBF); //wait for transmit buffer empty
-    U2TXREG = val & 0xff;
-    while (U2STAbits.UTXBF); //wait for transmit buffer empty
-    U2TXREG = (val >> 8)&0xff;
-}
-
-void sendAddress2(char address) {   //9-bit mode only
-    while (U2STAbits.UTXBF); //wait for transmit buffer empty
-    U2TXREG = 0x100 + address; // send the address with the 9th bit set
-}
-
-
-/*----UART 2 on SCL, SDA----------------*/
-void initUART2_passthrough(unsigned int BAUD) {
-    /*---------UART2 passthrough------------*/
-    configUART2(BAUD);
-    _U1RXIE =1;  //enable receive interrupt for uart1
-    _U2RXIE =1;  //enable receive interrupt for uart2
-    
-    DELAY_105uS
 
 }
 
@@ -1231,37 +1138,6 @@ void sqr4(uint16 w,uint16 R0,uint16 R1,uint16 RS1,uint16 R2,uint16 RS2,uint16 R3
 
 }
 
-void initUART(unsigned int BAUD) {
-    /*---------UART------------*/
-    TRISBbits.TRISB8 = 1; // B8 set as input(RX). connected to TX of MCP2200
-    ANSELBbits.ANSB8 = 0; // set B8 as digital input.
-    TRISBbits.TRISB7 = 0; // set as output. connected to RX of MCP2200
-
-    RPOR2bits.RP39R = 0x01; //Map B7(RP39) to UART TX
-    RPINR18bits.U1RXR = 0x28; //Map B8(RP40) to UART1 RX
-
-    U1MODEbits.STSEL = 0; //1 stop bit
-    U1MODEbits.PDSEL = 0; //no parity, 8-data bits
-    U1MODEbits.ABAUD = 0; //disable auto-baud
-    U1MODEbits.BRGH = 1; //high speed mode
-    U1BRG = BAUD;
-    U1MODEbits.UEN = 0;
-    U1MODEbits.RTSMD = 1;
-
-    U1STAbits.URXISEL = 0; //interrupt on 1 char recv
-
-    //IEC0bits.U1TXIE = 1; //enable TX interrupt
-
-    U1MODEbits.UARTEN = 1; //enable UART
-    U1STAbits.UTXEN = 1; //enable UART TX
-
-    U1MODEbits.URXINV = 0;
-
-    DELAY_105uS
-    while(hasChar())getChar(); //clear buffer
-
-}
-
 void setADCMode(BYTE mode,BYTE chosa,BYTE ch123sa){
     if(ADC_MODE == mode && chosa==CHOSA && ch123sa == CH123SA)return;
     else{
@@ -1604,55 +1480,6 @@ void read_flash(_prog_addressT pointer, BYTE location) {
     }
 }
 
-
-bool hasChar() {
-    return U1STAbits.URXDA;
-}
-
-void sendChar(BYTE val) {
-    while (U1STAbits.UTXBF); //wait for transmit buffer empty
-    U1TXREG = val;
-}
-
-void sendInt(unsigned int val) {
-    while (U1STAbits.UTXBF); //wait for transmit buffer empty
-    U1TXREG = val & 0xff;
-    while (U1STAbits.UTXBF); //wait for transmit buffer empty
-    U1TXREG = (val >> 8)&0xff;
-}
-
-void sendLong(unsigned int lsb,unsigned int msb) {
-    while (U1STAbits.UTXBF); //wait for transmit buffer empty
-    U1TXREG = lsb & 0xff;
-    while (U1STAbits.UTXBF); //wait for transmit buffer empty
-    U1TXREG = (lsb >> 8)&0xff;
-    while (U1STAbits.UTXBF); //wait for transmit buffer empty
-    U1TXREG = msb & 0xff;
-    while (U1STAbits.UTXBF); //wait for transmit buffer empty
-    U1TXREG = (msb >> 8)&0xff;
-}
-
-void ack(BYTE response) {
-    while (U1STAbits.UTXBF); //wait for transmit buffer empty
-    U1TXREG = response;
-}
-
-char getChar() {
-    while (!hasChar());
-    if (U1STAbits.FERR == 1) {
-        U1RXREG;
-        U1STAbits.OERR = 0;
-        return 0;
-    }
-    return U1RXREG;
-}
-
-unsigned int getInt() {
-    c1 = getChar()&0xFF;
-    c2 = getChar()&0xFF;
-    return (c2 << 8) | c1;
-}
-
 void initI2C(void) {
 
     _TRISB4 = 1; // set SCL and SDA pins as inputs.
@@ -1917,5 +1744,6 @@ void ReadPayload(BYTE num, BYTE* data)
     CSN_HIGH;
     WriteRegister(NRF_STATUS, 0x40);
 }
+
