+++ PSLab_Original/nbproject/private/private.xml
@@ -3,6 +3,10 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/1"/>
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group/>
+        <group>
+            <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/functions.c</file>
+            <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/proto2_main.c</file>
+            <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/PSLAB_UART.c</file>
+        </group>
     </open-files>
 </project-private>
