+++ PSLab_Original/proto2_main.c
@@ -11,11 +11,12 @@
 #pragma config ICS = NONE               // ICD Communication Channel Select bits (Reserved, do not use)
 #pragma config JTAGEN = OFF             // JTAG Enable bit (JTAG is disabled)
 
-#include<xc.h>
+#include <xc.h>
 #include <p24EP256GP204.h>
 #include <stdlib.h>
-#include<libpic30.h>
-#include"../PSLab_Original_Library/functions.c"
+#include <libpic30.h>
+#include "functions.c"
+#include "PSLAB_UART.c"
 //_FICD(ICS_PGD2 & JTAGEN_OFF) //Programmming pins  ..PGED2
