--- PSLab_Original_Library/nbproject/Makefile-genesis.properties
@@ -1,9 +1,0 @@
-#
-#Wed May 10 23:49:52 IST 2017
-default.com-microchip-mplab-nbide-toolchainXC16-XC16LanguageToolchain.md5=12db37d935262b866305dea1f2c952e5
-default.languagetoolchain.dir=/opt/microchip/xc16/v1.25/bin
-configurations-xml=76e7b1f74a3e0938d6c3acac37c2e11a
-com-microchip-mplab-nbide-embedded-makeproject-MakeProject.md5=f4f319408af34d2528a0ee2df2d21b1a
-default.languagetoolchain.version=1.25
-host.platform=linux
-conf.ids=default
