--- PSLab_Original_Library/nbproject/Makefile-variables.mk
@@ -1,13 +1,0 @@
-#
-# Generated - do not edit!
-#
-# NOCDDL
-#
-CND_BASEDIR=`pwd`
-# default configuration
-CND_ARTIFACT_DIR_default=dist/default/production
-CND_ARTIFACT_NAME_default=PSLab_Original_Library.a
-CND_ARTIFACT_PATH_default=dist/default/production/PSLab_Original_Library.a
-CND_PACKAGE_DIR_default=${CND_DISTDIR}/default/package
-CND_PACKAGE_NAME_default=PSLabOriginalLibrary.tar
-CND_PACKAGE_PATH_default=${CND_DISTDIR}/default/package/PSLabOriginalLibrary.tar
