--- PSLab_Original_Library/nbproject/private/configurations.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configurationDescriptor version="62">
-  <projectmakefile>Makefile</projectmakefile>
-  <defaultConf>0</defaultConf>
-  <confs>
-    <conf name="default" type="3">
-      <platformToolSN></platformToolSN>
-      <languageToolchainDir>/opt/microchip/xc16/v1.25/bin</languageToolchainDir>
-      <mdbdebugger version="1">
-        <placeholder1>place holder 1</placeholder1>
-        <placeholder2>place holder 2</placeholder2>
-      </mdbdebugger>
-      <runprofile version="6">
-        <args></args>
-        <rundir></rundir>
-        <buildfirst>true</buildfirst>
-        <console-type>0</console-type>
-        <terminal-type>0</terminal-type>
-        <remove-instrumentation>0</remove-instrumentation>
-        <environment>
-        </environment>
-      </runprofile>
-    </conf>
-  </confs>
-</configurationDescriptor>
