--- PSLab_Original_Library/nbproject/project.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://www.netbeans.org/ns/project/1">
-    <type>com.microchip.mplab.nbide.embedded.makeproject</type>
-    <configuration>
-        <data xmlns="http://www.netbeans.org/ns/make-project/1">
-            <name>PSLab_Original_Library</name>
-            <creation-uuid>612c9376-ea86-4599-b060-789d779fd9c6</creation-uuid>
-            <make-project-type>0</make-project-type>
-            <c-extensions>c</c-extensions>
-            <cpp-extensions/>
-            <header-extensions>h</header-extensions>
-            <asminc-extensions/>
-            <sourceEncoding>ISO-8859-1</sourceEncoding>
-            <make-dep-projects/>
-        </data>
-    </configuration>
-</project>
